ajustes, novas imagens e finalizado mil filmes

+++ android/app/src/main/java/com/tiagohs/cinema_history/models/tmdb/TMDBList.kt
@@ -12,7 +12,7 @@ data class TMDBList (
     @SerializedName("page") 			val page : Int,
     @SerializedName("iso_3166_1") 	val iso_3166_1 : String,
     @SerializedName("total_results") 	val total_results : Int,
-    @SerializedName("revenue") 		val revenue : Int,
+    @SerializedName("revenue") 		val revenue : Long,
     @SerializedName("total_pages") 	val total_pages : Int,
     @SerializedName("name") 			val name : String,
     @SerializedName("public") 		val public : Boolean,
@@ -21,6 +21,6 @@ data class TMDBList (
     @SerializedName("backdrop_path") 	val backdrop_path : String,
     @SerializedName("results") 		val results : List<Movie>,
     @SerializedName("average_rating") val average_rating : Double,
-    @SerializedName("runtime") 		val runtime : Int,
+    @SerializedName("runtime") 		val runtime : Long,
     @SerializedName("poster_path") 	val poster_path : String
 ): Serializable
